55 commits. CI on night-2 is green on all ten jobs, including E9 temporal and E15 frankenphp,
which were both red on main before this work.

What the cycle actually bought, in the order it matters:

**Four gates were found to be incapable of failing**, which is worse than a missing gate because
it reads as a pass. The PHP unit suite had never executed in CI (no unzip in the image, both
composer steps ended in `|| echo`, smoke printed "skipped" into a green run). E15 revolt had
reported IGNIS_PASSED=0 since the package split renamed test/ to tests/. The E9 negative control
printed REPLAY_OK while sdk-core evicted for nondeterminism. And the --all-features clippy step
added last night had never linted a line, because its job's toolchain ships without the component.

**A blocking quality gate from zero** (ADR-0041): clippy, rustfmt, cargo-deny, cargo-llvm-cov,
PHPStan 8, PHPUnit, php-cs-fixer. The debt was driven to zero in the same body of work rather than
deferred to a baseline file. Rust tests 9 -> 60 (63 with --all-features), coverage 7.67% -> 29.95%;
PHP tests 42 -> 209, coverage 5.61% -> 36.60% with a floor in CI; undocumented unsafe blocks
101 -> 0.

**Defects the new tests found**, three of them data-losing: pg::bind saturated an out-of-range
JSON number instead of refusing it; Proto::varint never terminated on a negative int; Proto::decode
invented a value for a truncated message off the network. Also the flat response header map that
dropped every Set-Cookie but the last, a streaming producer that was never told the client left,
and a blocking flock inside a fiber that took the OS thread down for good (V-81).

**Two owner decisions were overturned by measurement and not built**, both recorded with the
numbers: the file-session boot refusal (V-80) and step 2 of boundary typing.

**Numbers corrected rather than quietly dropped**: V-10's flatness claim reproduces, its absolutes
do not, and V-82 decomposes the RSS growth into ~10 MB outside this repository and 4.7 MB of ours.
The wrk-hello bisect is refused with a number instead of run, because the effect is smaller than
the instrument's noise.

+++ crates/ignis/src/php/embed.rs
@@ -28,6 +28,8 @@ struct Argv {
 }
 // SAFETY: written once before any thread reads it; the pointed-to strings are never freed.
 unsafe impl Sync for Argv {}
+// SAFETY: as above -- the pointers are process-lifetime and immutable after Engine::init, so moving
+// the struct between threads hands over nothing that can be freed or mutated.
 unsafe impl Send for Argv {}
 static ARGV: std::sync::OnceLock<Argv> = std::sync::OnceLock::new();
 
@@ -55,6 +57,9 @@ impl Engine {
         let ini = std::env::var("IGNIS_PHP_INI").ok().and_then(|p| CString::new(p).ok());
         // PHP_BINARY (E15a finding: empty under the stock embed, breaks suites that re-exec PHP).
         let exe = std::env::current_exe().ok().and_then(|p| CString::new(p.to_string_lossy().into_owned()).ok());
+        // SAFETY: single-threaded process startup, before any other thread exists. php_embed_module
+        // is libphp's own static and the hooks installed here are `extern "C"` functions with the
+        // signatures the SAPI struct declares; the argv pointers outlive the process.
         let rc = unsafe {
             sys::php_embed_module.startup = Some(module::ignis_sapi_startup);
             sys::php_embed_module.register_server_variables = Some(register_server_variables);
@@ -75,7 +80,7 @@ impl Engine {
             }
             sys::php_embed_init(argc, argv_ptrs.as_mut_ptr())
         };
-        if rc != sys::SUCCESS as i32 {
+        if rc != sys::SUCCESS {
             bail!("php_embed_init failed ({rc})");
         }
         #[cfg(php_async_abi)]
@@ -104,12 +109,16 @@ impl Engine {
         // cannot tell `exit(0)` from a parse error — both would look like "failed, status 0".
         // A sentinel in EG(exit_status) distinguishes them: only exit() overwrites it.
         unsafe { set_exit_status(SENTINEL) };
+        // SAFETY: PHP thread after startup. `code` and `name` are CStrings owned by this frame and
+        // outlive the call, and a null return-value slot means "discard the result", which is what
+        // php-cli's -r does.
         let rc = unsafe { sys::zend_eval_stringl_ex(code.as_ptr(), code.as_bytes().len(), std::ptr::null_mut(), name.as_ptr(), true) };
         // SAFETY: on a PHP thread after startup, same contract as run_file's use of it.
         let status = unsafe { exit_status() };
         if status != SENTINEL {
             return Ok(status); // exit(N) was called, N == 0 included
         }
+        // SAFETY: PHP thread after startup, the contract set_exit_status documents.
         unsafe { set_exit_status(0) };
         // A parse error or an uncaught throw: PHP has already printed it, and php-cli exits 255.
         if rc != sys::SUCCESS { Ok(255) } else { Ok(0) }
@@ -125,6 +134,8 @@ const SENTINEL: i32 = i32::MIN;
 /// # Safety
 /// PHP thread after startup; same contract as [`exit_status`].
 unsafe fn set_exit_status(v: i32) {
+    // SAFETY: the caller upholds `# Safety` above, so this thread has a TSRM context and the offset
+    // into it is the one the linked libphp reports; the field is a plain int owned by that thread.
     unsafe {
         let base = sys::tsrm_get_ls_cache() as *mut u8;
         let eg = base.add(sys::executor_globals_offset) as *mut sys::zend_executor_globals;
